[#184010483] Start view and delete

+++ AppJuegos/api/TicketConfiguration/TicketConfigurationApiviews.py
@@ -20,14 +20,20 @@ class TicketConfigurationViewSet(CRUDViewSet):
 
     def update(self, request, pk):
         ticket_configuration = TicketConfiguration.objects.get(id=pk)
+
         new_image = request.data.get('logo')
+        print(new_image)
         if new_image:
             serializer = TicketConfigurationSerializer(ticket_configuration, data=request.data)
             if serializer.is_valid():
-                old_image = ticket_configuration.logo
-                os.remove(old_image.path)
-                serializer.save()
-                return Response(serializer.data, status=status.HTTP_201_CREATED)
+                if ticket_configuration.logo == '':
+                    serializer.save()
+                    return Response(serializer.data, status=status.HTTP_201_CREATED)
+                else:
+                    old_image = ticket_configuration.logo
+                    os.remove(old_image.path)
+                    serializer.save()
+                    return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
             serializer = TicketConfigurationUpdateNotImageSerializer(ticket_configuration, data=request.data)
             if serializer.is_valid():
